Refactor request compression settings
