+++ src/endpoints/settings.js
@@ -4,6 +4,7 @@ import path from 'node:path';
 import express from 'express';
 import _ from 'lodash';
 import { sync as writeFileAtomicSync } from 'write-file-atomic';
+import bytes from 'bytes';
 
 import { SETTINGS_FILE } from '../constants.js';
 import { getConfigValue, generateTimestamp, removeOldBackups } from '../util.js';
@@ -13,7 +14,9 @@ import { getFileNameValidationFunction } from '../middleware/validateFileName.js
 const ENABLE_EXTENSIONS = !!getConfigValue('extensions.enabled', true, 'boolean');
 const ENABLE_EXTENSIONS_AUTO_UPDATE = !!getConfigValue('extensions.autoUpdate', true, 'boolean');
 const ENABLE_ACCOUNTS = !!getConfigValue('enableUserAccounts', false, 'boolean');
-const ENABLE_SAVE_UPLOAD_COMPRESSION = !!getConfigValue('performance.saveUploadCompression.enabled', true, 'boolean');
+const ENABLE_REQUEST_COMPRESSION = !!getConfigValue('performance.requestCompression.enabled', true, 'boolean');
+const REQUEST_COMPRESSION_THRESHOLD = bytes.parse(getConfigValue('performance.requestCompression.threshold', '256kb'));
+const REQUEST_COMPRESSION_TIMEOUT = getConfigValue('performance.requestCompression.timeout', 3000, 'number');
 
 // 10 minutes
 const AUTOSAVE_INTERVAL = 10 * 60 * 1000;
@@ -283,7 +286,11 @@ router.post('/get', (request, response) => {
         enable_extensions: ENABLE_EXTENSIONS,
         enable_extensions_auto_update: ENABLE_EXTENSIONS_AUTO_UPDATE,
         enable_accounts: ENABLE_ACCOUNTS,
-        enable_save_upload_compression: ENABLE_SAVE_UPLOAD_COMPRESSION,
+        request_compression: {
+            enabled: ENABLE_REQUEST_COMPRESSION,
+            threshold: REQUEST_COMPRESSION_THRESHOLD,
+            timeout: REQUEST_COMPRESSION_TIMEOUT,
+        },
     });
 });
